// src/app/views/dashboard/Setup/students/models/student.ts
export class Student {
  id?: number;
  nombres?: string;
  apellido_pa?: string;
  apellido_ma?: string;
  dni?: string;
  correo_electronico?: string;
  celular?: string;
  sexo?: string;
  fe_nacimiento?: string;
  direccion?: string;
  codigo?: string;
  departamento_id?: string;
  provincia_id?: string;
  distrito_id?: string;
  estado_civil_id?: number;
  grado_instruccion_id?: number;
}